코멘트 작성 삭제 검색기능 추가 및 BidId long으로 설정했던것 int로 수정

+++ BE/aivleTK/src/main/java/com/nara/aivleTK/service/bid/BidService.java
@@ -1,12 +1,12 @@
 package com.nara.aivleTK.service;
-import com.nara.aivleTK.dto.BidResponse;
+import com.nara.aivleTK.dto.bid.BidResponse;
 import java.util.List;
 
 public interface BidService {
     List<BidResponse> searchBid(String keyword);
     //나중에 투찰 마감시간이 지난 공고는 제외 할 수 있도록 변경
     List<BidResponse> getAllBid();
-    BidResponse getBidById(long id);
+    BidResponse getBidById(int id);
 
 
 }
